Move gloabl styles to styled comoonents
